+++ src/objnet/uartonbinterface.h
@@ -66,7 +66,7 @@ public:
     
     int availableWriteCount();
     
-    int addFilter(unsigned long id, unsigned long mask=0xFFFFFFFF);
+    int addFilter(uint32_t id, uint32_t mask=0xFFFFFFFF);
     void removeFilter(int number);
     
     bool busPresent();
